(f) some changes
